+++ Common.Mod.Common/Core/SystemSide.cs
@@ -1,7 +1,0 @@
-namespace Common.Mod.Common.Core;
-
-public enum SystemSide
-{
-    Client,
-    Server
-}